--- src/app/design/styled-button/styled-button.component.html
@@ -1,1 +1,0 @@
-<button [disabled]="disabled" (click)="onClick.emit($event)" [class]="flavor"><ng-content></ng-content></button>
