+++ src/core/components/button/button.component.html
@@ -1,0 +1,3 @@
+<button (click)="onClick.emit($event)" [class]="flavor" [disabled]="disabled">
+  <ng-content></ng-content>
+</button>
