+++ src/core/components/button/button.component.scss
@@ -37,6 +37,7 @@ button {
   border: none;
   font-size: 16px;
   margin-right: 10px;
+
   &:last-child {
     margin-right: 0;
   }
