+++ src/core/components/button/button.module.ts
@@ -1,0 +1,17 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {ButtonComponent} from './button.component';
+
+@NgModule({
+  declarations: [
+    ButtonComponent
+  ],
+  exports: [
+    ButtonComponent
+  ],
+  imports: [
+    CommonModule
+  ]
+})
+export class ButtonModule {
+}
